interface/requirements.txt
pyswip
spotipy
python-dotenv
